SonarQube linter

+++ src/main/java/com/sumdu/petrenko/diplom/controllers/FriendRequestController.java
@@ -2,6 +2,13 @@ package com.sumdu.petrenko.diplom.controllers;
 
 import com.sumdu.petrenko.diplom.models.FriendRequest;
 import com.sumdu.petrenko.diplom.services.FriendRequestService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,37 +19,112 @@ import java.util.Optional;
 
 @RestController
 @RequestMapping("/friendrequests")
+@Tag(name = "FriendRequests", description = "Operations related to friend requests")
 public class FriendRequestController {
+    private final FriendRequestService friendRequestService;
+
     @Autowired
-    private FriendRequestService friendRequestService;
+    public FriendRequestController(FriendRequestService friendRequestService) {
+        this.friendRequestService = friendRequestService;
+    }
 
     @GetMapping("/{id}")
-    public ResponseEntity<FriendRequest> getFriendRequestById(@PathVariable Long id) {
+    @Operation(
+            summary = "Get friend request by ID",
+            description = "Fetches a friend request by its unique ID.",
+            tags = {"Retrieve"},
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "Successfully retrieved the friend request",
+                            content = @Content(mediaType = "application/json", schema = @Schema(implementation = FriendRequest.class))),
+                    @ApiResponse(responseCode = "404", description = "Friend request not found", content = @Content)
+            }
+    )
+    public ResponseEntity<FriendRequest> getFriendRequestById(
+            @Parameter(description = "ID of the friend request to be fetched") @PathVariable Long id) {
         Optional<FriendRequest> friendRequest = friendRequestService.getFriendRequestById(id);
+
         return friendRequest.map(value -> new ResponseEntity<>(value, HttpStatus.OK))
                 .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @GetMapping("/sender/{senderId}")
-    public ResponseEntity<List<FriendRequest>> getFriendRequestsBySenderId(@PathVariable Long senderId) {
+    @Operation(
+            summary = "Get friend requests by sender ID",
+            description = "Fetches a list of friend requests sent by a specific user.",
+            tags = {"Retrieve"},
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "Successfully retrieved friend requests list",
+                            content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = FriendRequest.class)))),
+                    @ApiResponse(responseCode = "404", description = "Friend requests not found", content = @Content)
+            }
+    )
+    public ResponseEntity<List<FriendRequest>> getFriendRequestsBySenderId(
+            @Parameter(description = "ID of the sender whose friend requests are to be fetched") @PathVariable Long senderId) {
         List<FriendRequest> friendRequests = friendRequestService.getFriendRequestsBySenderId(senderId);
+
+        if (friendRequests.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         return new ResponseEntity<>(friendRequests, HttpStatus.OK);
     }
 
     @GetMapping("/receiver/{receiverId}")
-    public ResponseEntity<List<FriendRequest>> getFriendRequestsByReceiverId(@PathVariable Long receiverId) {
+    @Operation(
+            summary = "Get friend requests by receiver ID",
+            description = "Fetches a list of friend requests received by a specific user.",
+            tags = {"Retrieve"},
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "Successfully retrieved friend requests list",
+                            content = @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = FriendRequest.class)))),
+                    @ApiResponse(responseCode = "404", description = "Friend requests not found", content = @Content)
+            }
+    )
+    public ResponseEntity<List<FriendRequest>> getFriendRequestsByReceiverId(
+            @Parameter(description = "ID of the receiver whose friend requests are to be fetched") @PathVariable Long receiverId) {
         List<FriendRequest> friendRequests = friendRequestService.getFriendRequestsByReceiverId(receiverId);
+
+        if (friendRequests.isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
         return new ResponseEntity<>(friendRequests, HttpStatus.OK);
     }
 
     @PostMapping
+    @Operation(
+            summary = "Create a new friend request",
+            description = "Creates a new friend request with the provided details.",
+            tags = {"Create"},
+            responses = {
+                    @ApiResponse(responseCode = "201", description = "Friend request created successfully", content = @Content),
+                    @ApiResponse(responseCode = "400", description = "Invalid friend request data", content = @Content)
+            }
+    )
     public ResponseEntity<Void> createFriendRequest(@RequestBody FriendRequest friendRequest) {
         friendRequestService.saveFriendRequest(friendRequest);
+
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteFriendRequest(@PathVariable Long id) {
+    @Operation(
+            summary = "Delete a friend request by ID",
+            description = "Deletes a friend request with the specified ID.",
+            tags = {"Delete"},
+            responses = {
+                    @ApiResponse(responseCode = "204", description = "Friend request successfully deleted", content = @Content),
+                    @ApiResponse(responseCode = "404", description = "Friend request not found", content = @Content)
+            }
+    )
+    public ResponseEntity<Void> deleteFriendRequest(
+            @Parameter(description = "ID of the friend request to be deleted") @PathVariable Long id) {
+        Optional<FriendRequest> friendRequestOptional = friendRequestService.getFriendRequestById(id);
+
+        if (friendRequestOptional.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         friendRequestService.deleteFriendRequest(id);
         return ResponseEntity.noContent().build();
     }
